refactor(outputgraph): グラフ１つを出力する処理を関数化
